moved onTap to color circle
